+++ src-core/modules/ccsds/module_ccsds_simple_psk_decoder.h
@@ -64,8 +64,8 @@ namespace ccsds
         std::ifstream data_in;
         std::ofstream data_out;
 
-        std::atomic<size_t> filesize;
-        std::atomic<size_t> progress;
+        std::atomic<uint64_t> filesize;
+        std::atomic<uint64_t> progress;
 
         std::shared_ptr<deframing::BPSK_CCSDS_Deframer> deframer;
         std::shared_ptr<deframing::BPSK_CCSDS_Deframer> deframer_qpsk;
